+++ runway/__init__.py
@@ -1,2 +1,2 @@
 """Set package version."""
-__version__ = '0.9.1'
+__version__ = '0.10.0rc1'