+++ runway/cli.py
@@ -5,6 +5,7 @@ Usage:
   runway (test|preflight)
   runway (plan|taxi)
   runway (deploy|takeoff)
+  runway (destroy|dismantle)
   runway gitclean
   runway gen-sample (cfn|sls|stacker|tf)
   runway -h | --help